+++ packages/vant2/docs/demos/guide/field/template.vue
@@ -1,15 +1,15 @@
 <template>
   <FormProvider :form="form">
     <Field
-      name="field"
+      name="input"
       title="输入框"
-      :component="[VanField, { label: '输入框', placeholder: '请输入' }]"
+      :component="[Input, { label: '输入框', placeholder: '请输入' }]"
     />
     <Field
       name="textarea"
       title="文本框"
       :component="[
-        VanField,
+        Input,
         {
           autosize: true,
           type: 'textarea',
@@ -19,10 +19,10 @@
       ]"
     />
     <Field
-      name="fieldRead"
+      name="inputRead"
       title="输入框"
       pattern="readPretty"
-      :component="[VanField, { label: '输入框', placeholder: '请输入' }]"
+      :component="[Input, { label: '输入框', placeholder: '请输入' }]"
     />
     <Submit :style="{ 'margin-top': '16px' }" round block @submit="log">
       提交
@@ -33,7 +33,7 @@
 <script>
 import { createForm } from '@formily/core'
 import { FormProvider, Field } from '@formily/vue'
-import { Field as VanField, Submit } from '@formily/vant2'
+import { Input, Submit } from '@formily/vant2'
 
 const form = createForm()
 
@@ -41,7 +41,7 @@ export default {
   components: { FormProvider, Field, Submit },
   data() {
     return {
-      VanField,
+      Input,
       form,
     }
   },
