+++ packages/vant2/docs/demos/guide/form/markup-schema.vue
@@ -2,14 +2,14 @@
   <Form :form="form" @autoSubmit="log" @autoSubmitFailed="log">
     <SchemaField>
       <SchemaStringField
-        name="field"
-        x-component="Field"
+        name="input"
+        x-component="Input"
         :x-component-props="{ label: '输入框', placeholder: '请输入' }"
       />
 
       <SchemaStringField
         name="switch"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '开关',
         }"
@@ -21,7 +21,7 @@
 
       <SchemaStringField
         name="checkbox"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '复选框',
         }"
@@ -33,7 +33,7 @@
 
       <SchemaArrayField
         name="checkboxGroup"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '复选框组',
         }"
@@ -49,7 +49,7 @@
 
       <SchemaArrayField
         name="radio"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '单选框',
         }"
@@ -65,7 +65,7 @@
 
       <SchemaStringField
         name="stepper"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '步进器',
         }"
@@ -74,7 +74,7 @@
 
       <SchemaStringField
         name="rate"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '评分',
         }"
@@ -83,7 +83,7 @@
 
       <SchemaStringField
         name="slider"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '滑块',
         }"
@@ -92,7 +92,7 @@
 
       <SchemaStringField
         name="uploader"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '文件上传',
         }"
@@ -202,7 +202,8 @@ import { createForm } from '@formily/core'
 import { createSchemaField } from '@formily/vue'
 import {
   Form,
-  Field,
+  FormItem,
+  Input,
   Switch,
   Checkbox,
   Radio,
@@ -220,7 +221,8 @@ import {
 const form = createForm()
 const fields = createSchemaField({
   components: {
-    Field,
+    Input,
+    FormItem,
     Switch,
     Checkbox,
     Radio,
