+++ packages/vant2/docs/demos/guide/preview-text/base.vue
@@ -2,15 +2,15 @@
   <Form :form="form">
     <SchemaField>
       <SchemaStringField
-        name="field"
-        x-component="PreviewText.Field"
+        name="input"
+        x-component="PreviewText.Input"
         :x-component-props="{ label: '文本预览' }"
         default="Hello world"
       />
 
       <SchemaStringField
         name="switch"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{ label: '开关' }"
         x-component="PreviewText.Switch"
         :default="true"
@@ -18,7 +18,7 @@
 
       <SchemaArrayField
         name="checkboxGroup"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '复选框组',
         }"
@@ -36,7 +36,7 @@
 
       <SchemaArrayField
         name="radio"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '单选框',
         }"
@@ -53,7 +53,7 @@
 
       <SchemaStringField
         name="stepper"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '步进器',
         }"
@@ -63,7 +63,7 @@
 
       <SchemaStringField
         name="rate"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '评分',
         }"
@@ -76,7 +76,7 @@
 
       <SchemaStringField
         name="slider"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '滑块',
         }"
@@ -86,7 +86,7 @@
 
       <SchemaStringField
         name="uploader"
-        x-decorator="Field"
+        x-decorator="FormItem"
         :x-decorator-props="{
           label: '文件上传',
         }"
@@ -200,13 +200,13 @@
 <script>
 import { createForm } from '@formily/core'
 import { createSchemaField } from '@formily/vue'
-import { Form, PreviewText, Field, Submit } from '@formily/vant2'
+import { Form, PreviewText, FormItem, Submit } from '@formily/vant2'
 
 const form = createForm()
 const fields = createSchemaField({
   components: {
     PreviewText,
-    Field,
+    FormItem,
   },
 })
