+++ packages/vant2/src/form-item/style.ts
@@ -1,0 +1,5 @@
+import 'vant/lib/field/index.less'
+
+// 依赖
+import '../preview-text/style'
+import '../input/style.scss'
