+++ packages/vant2/src/index.ts
@@ -6,9 +6,10 @@ export * from './cascader'
 export * from './checkbox'
 export * from './datetime-picker'
 export * from './form'
+export * from './form-item'
 export * from './form-layout'
 export * from './picker'
-export * from './field'
+export * from './input'
 export * from './preview-text'
 export * from './radio'
 export * from './rate'
