+++ packages/vant2/src/preview-text/style.scss
@@ -1,6 +1,6 @@
 @import '../__builtins__/styles/common.scss';
 
-.#{$formily-prefix}-preview-field {
+.#{$formily-prefix}-preview-input {
   position: relative;
   display: -webkit-box;
   display: -webkit-flex;
@@ -15,7 +15,7 @@
   background-color: #fff;
 }
 
-.#{$formily-prefix}-preview-field-title {
+.#{$formily-prefix}-preview-input-title {
   display: inline-block;
   width: 6.2em;
   color: #646566;
