+++ packages/vant2/src/style.ts
@@ -3,4 +3,4 @@ import './preview-text/style.scss'
 import './form-layout/style.scss'
 import './form/style.scss'
 import './space/style.scss'
-import './field/style.scss'
+import './input/style.scss'
